+++ src/gui/ListCtrl.h
@@ -37,9 +37,9 @@ public:
     void UpdateSelected();
     
     // Getters/Setters
-    int GetColumnClicked() const                    { return m_column; }
-    int GetItem( long index ) const;
-    int GetItemCount() const                        { return wxListCtrl::GetItemCount(); }
+    int  GetColumnClicked() const                   { return m_column; }
+    long GetItem( long index ) const;
+    int  GetItemCount() const                       { return wxListCtrl::GetItemCount(); }
     bool SetColumnWidth( int col, int width )       { return wxListCtrl::SetColumnWidth( col, width ); }
     void SetMaxSize( const wxSize &size )           { wxListCtrl::SetMaxSize( size ); }
     void SetMinSize( const wxSize &size )           { wxListCtrl::SetMinSize( size ); }
